Error in ListView
There is an error in ListView on line 128 where it cannot read the id of undefied.

I tried multiple solutions but feel like I am overthinking it.

I also made a seperate dir and file for the API but since I included it in the .gitignore, I will send a screenshot just in case that turns out to be the problem.